Pull out smoke tests from `install_test.go`
As part of `install_test.go`, we have a couple of test cases meant to
act as "smoke tests" and flag any obvious failures before the rest of
the tests are run. The smoke tests are heavily coupled to the install
stages and are unnecessarily run for each suite. This adds a bit of an
overhead and makes it harder to change and manage the code in
`install_test.go`.

To fix, we split up the tests from the install tests. During the
process, some golden files had to be copied over into a new testdata
directory. There are also duplicated functions across the two which will
be refactored gradually. For this change, `getCheckOutput` has also been
refactored into its own method on the TestHelper struct.

At the moment, the tests will _not_ be run right after `install_test.go`
and will instead be ran as part of the deep suite. This is only
temporary, a more deterministic order will be enforced once more tests
are refactored.
