+++ test/integration/install_test.go
@@ -57,28 +57,6 @@ var (
 		{Namespace: "linkerd-multicluster", Name: "linkerd-gateway"},
 	}
 
-	injectionCases = []struct {
-		ns          string
-		annotations map[string]string
-		injectArgs  []string
-	}{
-		{
-			ns: "smoke-test",
-			annotations: map[string]string{
-				k8s.ProxyInjectAnnotation: k8s.ProxyInjectEnabled,
-			},
-			injectArgs: nil,
-		},
-		{
-			ns:         "smoke-test-manual",
-			injectArgs: []string{"--manual"},
-		},
-		{
-			ns:         "smoke-test-ann",
-			injectArgs: []string{},
-		},
-	}
-
 	//skippedInboundPorts lists some ports to be marked as skipped, which will
 	// be verified in test/integration/inject
 	skippedInboundPorts       = "1234,5678"
@@ -933,7 +911,7 @@ func testCheckCommand(t *testing.T, stage, expectedVersion, namespace, cliVersio
 		}
 	}
 
-	expected := getCheckOutput(t, golden, TestHelper.GetLinkerdNamespace())
+	expected := TestHelper.GetCheckOutput(t, golden, TestHelper.GetLinkerdNamespace())
 	timeout := time.Minute * 5
 	err := TestHelper.RetryFor(timeout, func() error {
 		if cliVersionOverride != "" {
@@ -955,20 +933,20 @@ func testCheckCommand(t *testing.T, stage, expectedVersion, namespace, cliVersio
 			if ext == multiclusterExtensionName {
 				// multicluster check --proxy and multicluster check have the same output
 				// so use the same golden file.
-				expected = getCheckOutput(t, "check.multicluster.golden", TestHelper.GetMulticlusterNamespace())
+				expected = TestHelper.GetCheckOutput(t, "check.multicluster.golden", TestHelper.GetMulticlusterNamespace())
 				if !strings.Contains(out, expected) {
 					return fmt.Errorf(
 						"Expected:\n%s\nActual:\n%s", expected, out)
 				}
 			} else if ext == vizExtensionName {
 				if stage == proxyStage {
-					expected = getCheckOutput(t, "check.viz.proxy.golden", TestHelper.GetVizNamespace())
+					expected = TestHelper.GetCheckOutput(t, "check.viz.proxy.golden", TestHelper.GetVizNamespace())
 					if !strings.Contains(out, expected) {
 						return fmt.Errorf(
 							"Expected:\n%s\nActual:\n%s", expected, out)
 					}
 				} else {
-					expected = getCheckOutput(t, "check.viz.golden", TestHelper.GetVizNamespace())
+					expected = TestHelper.GetCheckOutput(t, "check.viz.golden", TestHelper.GetVizNamespace())
 					if !strings.Contains(out, expected) {
 						return fmt.Errorf(
 							"Expected:\n%s\nActual:\n%s", expected, out)
@@ -982,35 +960,6 @@ func testCheckCommand(t *testing.T, stage, expectedVersion, namespace, cliVersio
 	if err != nil {
 		testutil.AnnotatedFatal(t, fmt.Sprintf("'linkerd check' command timed-out (%s)", timeout), err)
 	}
-}
-
-func getCheckOutput(t *testing.T, goldenFile string, namespace string) string {
-	pods, err := TestHelper.KubernetesHelper.GetPods(context.Background(), namespace, nil)
-	if err != nil {
-		testutil.AnnotatedFatal(t, fmt.Sprintf("failed to retrieve pods: %s", err), err)
-	}
-
-	proxyVersionErr := ""
-	err = healthcheck.CheckProxyVersionsUpToDate(pods, version.Channels{})
-	if err != nil {
-		proxyVersionErr = err.Error()
-	}
-
-	tpl := template.Must(template.ParseFiles("testdata" + "/" + goldenFile))
-	vars := struct {
-		ProxyVersionErr string
-		HintURL         string
-	}{
-		proxyVersionErr,
-		healthcheck.HintBaseURL(TestHelper.GetVersion()),
-	}
-
-	var expected bytes.Buffer
-	if err := tpl.Execute(&expected, vars); err != nil {
-		testutil.AnnotatedFatal(t, fmt.Sprintf("failed to parse check.viz.golden template: %s", err), err)
-	}
-
-	return expected.String()
 }
 
 // TODO: run this after a `linkerd install config`
@@ -1115,124 +1064,6 @@ func TestDashboard(t *testing.T) {
 	}
 }
 
-func TestInject(t *testing.T) {
-	resources, err := testutil.ReadFile("testdata/smoke_test.yaml")
-	if err != nil {
-		testutil.AnnotatedFatalf(t, "failed to read smoke test file",
-			"failed to read smoke test file: %s", err)
-	}
-
-	ctx := context.Background()
-	for _, tc := range injectionCases {
-		tc := tc // pin
-		t.Run(tc.ns, func(t *testing.T) {
-			var out string
-
-			prefixedNs := TestHelper.GetTestNamespace(tc.ns)
-
-			err := TestHelper.CreateDataPlaneNamespaceIfNotExists(ctx, prefixedNs, tc.annotations)
-			if err != nil {
-				testutil.AnnotatedFatalf(t, fmt.Sprintf("failed to create %s namespace", prefixedNs),
-					"failed to create %s namespace: %s", prefixedNs, err)
-			}
-
-			if tc.injectArgs != nil {
-				cmd := []string{"inject"}
-				cmd = append(cmd, tc.injectArgs...)
-				cmd = append(cmd, "testdata/smoke_test.yaml")
-
-				var injectReport string
-				out, injectReport, err = TestHelper.PipeToLinkerdRun("", cmd...)
-				if err != nil {
-					testutil.AnnotatedFatalf(t, "'linkerd inject' command failed",
-						"'linkerd inject' command failed: %s\n%s", err, out)
-				}
-
-				err = TestHelper.ValidateOutput(injectReport, "inject.report.golden")
-				if err != nil {
-					testutil.AnnotatedFatalf(t, "received unexpected output",
-						"received unexpected output\n%s", err.Error())
-				}
-			} else {
-				out = resources
-			}
-
-			out, err = TestHelper.KubectlApply(out, prefixedNs)
-			if err != nil {
-				testutil.AnnotatedFatalf(t, "'kubectl apply' command failed",
-					"'kubectl apply' command failed\n%s", out)
-			}
-
-			for _, deploy := range []string{"smoke-test-terminus", "smoke-test-gateway"} {
-				if err := TestHelper.CheckPods(ctx, prefixedNs, deploy, 1); err != nil {
-					if rce, ok := err.(*testutil.RestartCountError); ok {
-						testutil.AnnotatedWarn(t, "CheckPods timed-out", rce)
-					} else {
-						testutil.AnnotatedFatal(t, "CheckPods timed-out", err)
-					}
-				}
-			}
-
-			url, err := TestHelper.URLFor(ctx, prefixedNs, "smoke-test-gateway", 8080)
-			if err != nil {
-				testutil.AnnotatedFatalf(t, "failed to get URL",
-					"failed to get URL: %s", err)
-			}
-
-			output, err := TestHelper.HTTPGetURL(url)
-			if err != nil {
-				testutil.AnnotatedFatalf(t, "unexpected error",
-					"unexpected error: %v %s", err, output)
-			}
-
-			expectedStringInPayload := "\"payload\":\"BANANA\""
-			if !strings.Contains(output, expectedStringInPayload) {
-				testutil.AnnotatedFatalf(t, "application response doesn't contain the expected response",
-					"expected application response to contain string [%s], but it was [%s]",
-					expectedStringInPayload, output)
-			}
-		})
-	}
-}
-
-func TestServiceProfileDeploy(t *testing.T) {
-	bbProto, err := TestHelper.HTTPGetURL("https://raw.githubusercontent.com/BuoyantIO/bb/v0.0.5/api.proto")
-	if err != nil {
-		testutil.AnnotatedFatalf(t, "unexpected error",
-			"unexpected error: %v %s", err, bbProto)
-	}
-
-	for _, tc := range injectionCases {
-		tc := tc // pin
-		t.Run(tc.ns, func(t *testing.T) {
-			prefixedNs := TestHelper.GetTestNamespace(tc.ns)
-
-			cmd := []string{"profile", "-n", prefixedNs, "--proto", "-", "smoke-test-terminus-svc"}
-			bbSP, stderr, err := TestHelper.PipeToLinkerdRun(bbProto, cmd...)
-			if err != nil {
-				testutil.AnnotatedFatalf(t, "unexpected error",
-					"unexpected error: %v %s", err, stderr)
-			}
-
-			out, err := TestHelper.KubectlApply(bbSP, prefixedNs)
-			if err != nil {
-				testutil.AnnotatedFatalf(t, "'kubectl apply' command failed",
-					"'kubectl apply' command failed: %s\n%s", err, out)
-			}
-		})
-	}
-}
-
-func TestCheckProxy(t *testing.T) {
-	for _, tc := range injectionCases {
-		tc := tc // pin
-		t.Run(tc.ns, func(t *testing.T) {
-			prefixedNs := TestHelper.GetTestNamespace(tc.ns)
-			testCheckCommand(t, "proxy", TestHelper.GetVersion(), prefixedNs, "")
-		})
-	}
-}
-
 func TestRestarts(t *testing.T) {
 	expectedDeployments := testutil.LinkerdDeployReplicasEdge
 	if !TestHelper.ExternalPrometheus() {
@@ -1246,20 +1077,6 @@ func TestRestarts(t *testing.T) {
 				testutil.AnnotatedFatal(t, "CheckPods timed-out", err)
 			}
 		}
-	}
-}
-
-// TestCleanUp deletes the resources used in the above tests
-func TestCleanUp(t *testing.T) {
-	ctx := context.Background()
-	for _, tc := range injectionCases {
-		tc := tc // pin
-		t.Run(tc.ns, func(t *testing.T) {
-			prefixedNs := TestHelper.GetTestNamespace(tc.ns)
-			if err := TestHelper.DeleteNamespaceIfExists(ctx, prefixedNs); err != nil {
-				testutil.AnnotatedFatal(t, "Deleting namespace failed", err)
-			}
-		})
 	}
 }
