+++ test/integration/smoke/testdata/smoke_test.yaml
